% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{boot_control}
\alias{boot_control}
\title{Bootstrap control list}
\usage{
boot_control(nboot = NULL, seed = NULL, ncores = 1L)
}
\arguments{
\item{nboot}{Number of bootstrap replicates to produce.}

\item{seed}{Seed used to initialize parallel processing-safe random number
generation.}

\item{ncores}{Number of cores to use in parallel processing. Defaults to 1.}
}
\description{
Bootstrap control list
}
\details{
Can be used as an argument to the \code{control} argument in \code{crosswalk()},
but intended primarily to validate the list provided by that argument.
}
