% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{percentile_bootstrap_ci}
\alias{percentile_bootstrap_ci}
\title{Percentile method}
\usage{
percentile_bootstrap_ci(bootdist, alpha = 0.05)
}
\arguments{
\item{bootdist}{A vector containing the bootstrap distribution}

\item{alpha}{Alpha level to use for confidence interval calculation}
}
\description{
Percentile method
}
