% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{bootstrap_crosswalk}
\alias{bootstrap_crosswalk}
\title{Bootstrap a split routine}
\usage{
bootstrap_crosswalk(..., nboot, ncores = 1L, seed)
}
\arguments{
\item{...}{Pass arguments to \code{crosswalk()}}

\item{nboot}{Number of bootstrap replicates to generate}

\item{ncores}{Number of cores to use in parallel processing}

\item{seed}{Seed used by doRNG to generate reproducible parallel computations}

\item{alpha}{Alpha level for bootstrap confidence intervals}
}
\description{
Bootstrap a split routine
}
