% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{bootstrap_splits}
\alias{bootstrap_splits}
\title{Bootstrap a split routine}
\usage{
bootstrap_splits(..., num_boot, num_cores, rng_seed)
}
\arguments{
\item{...}{Pass arguments to \code{analyze_splits()}}

\item{num_boot}{Number of bootstrap replicates to generate}

\item{num_cores}{Number of cores to use in parallel processing}

\item{rng_seed}{Seed used by doRNG to generate reproducible parallel computations}
}
\description{
Bootstrap a split routine
}
