% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cogsim}
\alias{cogsim}
\title{Simulated cognitive data}
\format{
\subsection{\code{cogsim}}{

A data.table containing 1,557 rows and 4 columns:
\describe{
\item{subid}{Integer. Subject ID}
\item{dementia}{Integer. Binary indicator of dementia}
\item{mmse}{Integer. Analagous to MMSE_bl from ADNIMERGE}
\item{moca}{Integer. Analagous to MOCA_bl from ADNIMERGE}
}
}
}
\source{
\url{https://adni.loni.usc.edu}
}
\usage{
cogsim
}
\description{
A simulated dataset based on a subset of the ADNIMERGE dataset from the
Alzheimer's Disease Neuroimaging Initiative (ADNI)
}
\keyword{datasets}
