% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{do_crosswalk}
\alias{do_crosswalk}
\title{Crosswalk an effect estimate}
\usage{
do_crosswalk(
  object,
  est_mean = NULL,
  est_se = NULL,
  est_ci = NULL,
  est_alpha = 0.05,
  alpha = 0.05,
  est_indep = NULL,
  est_outcome = NULL
)
}
\arguments{
\item{object}{An object of class \code{cogxwalkr} or the result of \code{\link[=est_cw_coef]{est_cw_coef()}}}

\item{est_mean}{Point estimate (beta) to be crosswalked to the alternative outcome
measure}

\item{est_se}{The standard error of \code{est_mean}}

\item{est_ci}{The lower (1-alpha)\% confidence interval of \code{est_mean}}

\item{est_alpha}{The alpha level for the confidence interval (if \code{est_se} is provided)
or the alpha level that will be used to back-calculate the standard error from
\code{est_ci}. Defaults to 0.05.}

\item{alpha}{The alpha level for the confidence interval of the crosswalked estimate.
Defaults to 0.05.}

\item{est_indep}{The independent variable to which \code{est_mean} applies}

\item{est_outcome}{The outcome measure in the original study (e.g., "MOCA", "MMSE")}
}
\description{
Crosswalk an effect estimate
}
