% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{plot.cogxwalkr}
\alias{plot.cogxwalkr}
\title{Plot information about the bootstrap distribution}
\usage{
\method{plot}{cogxwalkr}(
  cx,
  which = c("boot", "slope"),
  breaks = "FD",
  sarg = list(col = "black", lty = 1, lwd = 2),
  barg = list(col = "red", lty = 2, lwd = 2)
)
}
\arguments{
\item{cx}{An object of class "cogxwalkr", i.e., an object returned by \code{\link[=crosswalk]{crosswalk()}}.}

\item{breaks}{Passed to \code{\link[graphics:hist]{graphics::hist()}}, overriding the default method with "FD"
(\code{\link[grDevices:nclass]{grDevices::nclass.FD()}}).}

\item{sarg}{List of parameters passed to the \code{\link[graphics:abline]{graphics::abline()}} that plots the sample
coefficient estimate}

\item{barg}{List of parameters passed to the \code{\link[graphics:abline]{graphics::abline()}} that plots the mean
coefficient estimate across bootstrap replicates}
}
\description{
Plot information about the bootstrap distribution
}
