% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{print.cogxwalkr.crosswalk}
\alias{print.cogxwalkr.crosswalk}
\title{Print the results of a crosswalk}
\usage{
\method{print}{cogxwalkr.crosswalk}(x, ..., digits = 3L)
}
\arguments{
\item{x}{An object of class "cogxwalkr.crosswalk", i.e., as returned by
\code{\link[=do_crosswalk]{do_crosswalk()}}}

\item{...}{Unused}

\item{digits}{Number of digits to print. Passed to the \code{digits} and \code{nsmall} arguments
to \code{\link[base:format]{base::format()}}.}
}
\description{
Print the results of a crosswalk
}
