% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{print.summary.cogxwalkr}
\alias{print.summary.cogxwalkr}
\title{Print a cogxwalkr summary}
\usage{
\method{print}{summary.cogxwalkr}(x, ..., digits = 3L)
}
\arguments{
\item{x}{An object of class "summary.cogxwalkr", i.e., as returned by
\code{\link[=summary.cogxwalkr]{summary.cogxwalkr()}}.}

\item{...}{Unused}

\item{digits}{Number of digits to print. Passed to the \code{digits} and \code{nsmall} arguments
to \code{\link[base:format]{base::format()}}.}
}
\description{
Print a cogxwalkr summary
}
