% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.r
\name{make_unconditional_splits}
\alias{make_unconditional_splits}
\alias{make_conditional_splits}
\alias{make_splits}
\title{Make an unconditional split dataset}
\usage{
make_unconditional_splits(data, niter)

make_conditional_splits(cdvar = NULL, data, loop = FALSE)

make_splits(cdvar = NULL, data, cdloop = FALSE, niter = NULL)
}
\arguments{
\item{data}{Input dataset}

\item{niter}{Number of split iterations to conduct}

\item{cdvar}{Character string naming auxiliary variable by which to condition splits}

\item{loop}{Boolean declaring whether to use a for loop. The default FALSE will
generate splits and operate on an expanded data.table in-memory. If your machine has
limited memory, set this argument to TRUE in order to process splits sequentially.
The default option should be much faster.}

\item{cdloop}{Boolean passed to \code{loop} argument of \code{make_conditional_splits()}. Ignored
when conducting unconditional splits.}
}
\description{
Make an unconditional split dataset

Make a conditional split dataset

Make a split dataset
}
