% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{summary.cogxwalkr}
\alias{summary.cogxwalkr}
\title{Summarize a cogxwalkr list}
\usage{
\method{summary}{cogxwalkr}(object, ..., alpha = 0.05, bci_type = c("percentile", "normal"))
}
\arguments{
\item{object}{An object of class "cogxwalkr", i.e., as returned by \code{\link[=crosswalk]{crosswalk()}}.}

\item{...}{Unused}

\item{alpha}{Alpha to use for confidence interval calculation. Defaults to 0.05.}

\item{bci_type}{Type of bootstrapped confidence interval to calculate. Currently
accepts "percentile" (or "perc") and/or "normal." Both are included by default.}
}
\description{
Summarize a cogxwalkr list
}
