% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Percentile method}
\usage{
bootstrap_ci(cx, alpha = 0.05, type = c("percentile", "normal"))
}
\arguments{
\item{cx}{An object of class \code{cogxwalkr}}

\item{alpha}{Alpha level to use for confidence interval calculation}
}
\description{
Percentile method
}
