% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{.percentile_bootstrap_ci}
\alias{.percentile_bootstrap_ci}
\title{Percentile method}
\usage{
.percentile_bootstrap_ci(bootdist, alpha = 0.05, sample_est = NULL)
}
\arguments{
\item{bootdist}{A vector containing the bootstrap distribution}

\item{alpha}{Alpha level to use for confidence interval calculation}

\item{sample_est}{The coefficient from a split procedure conducted in the original
sample}
}
\description{
Percentile method
}
