% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Calculate bootstrapped confidence limits}
\usage{
bootstrap_ci(cx, alpha = 0.05, type = "percentile")
}
\arguments{
\item{cx}{An object of class "cogxwalkr", i.e., an object returned by \code{crosswalk()}.}

\item{alpha}{Alpha level to use for confidence interval calculation}

\item{type}{Type of confidence limits to calculate. Currently supported:
percentile, normal.}
}
\description{
Calculate bootstrapped confidence limits
}
