% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{crosswalk}
\alias{crosswalk}
\title{Generate data splits}
\usage{
crosswalk(
  cog1,
  cog2,
  data,
  num_iter,
  type = "unconditional",
  condition_by = NULL,
  boot_ci = FALSE,
  boot_control = list(...)
)
}
\arguments{
\item{cog1}{The name of the first cognitive measure column}

\item{cog2}{The name of the second cognitive measure column}

\item{data}{A data.table or data.frame containing the cognitive measure data}

\item{num_iter}{Number of split iterations to conduct}

\item{type}{One of "unconditional" or "conditional". A conditional split will conduct
the splits independently within levels of an auxiliary variable (see \code{condition_by}).}

\item{condition_by}{The name of a conditioning variable by which splits will be
conducted when type="conditional"}

\item{boot_ci}{Boolean indicating whether to generate bootstrap confidence intervals}

\item{boot_control}{A list of settings passed to \code{bootstrap_crosswalk()}}
}
\description{
Generate data splits
}
