% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{crosswalk}
\alias{crosswalk}
\alias{est_cw_coef}
\title{Crosswalk cognitive measures}
\usage{
crosswalk(
  cog1,
  cog2,
  data,
  niter = NULL,
  condition_by = NULL,
  condition_loop = FALSE,
  control = list(...)
)

est_cw_coef(cog1, cog2, data, method = "lm")
}
\arguments{
\item{cog1}{The name of the first cognitive measure column}

\item{cog2}{The name of the second cognitive measure column}

\item{data}{A data.table, data.frame, matrix, or list containing the
cognitive measure data}

\item{niter}{Number of iterations to conduct for an unconditional split routine}

\item{condition_by}{The name of a conditioning variable by which splits will be
conducted. If not conducted, the function will use unconditional splits.}

\item{condition_loop}{Whether to conduct conditional splitting sequentially. Defaults
to FALSE to maximize speed. Unused if \code{condition_by} is NULL. See documentation
for \code{\link[=make_conditional_splits]{make_conditional_splits()}} for details.}

\item{control}{A list of settings passed to \code{\link[=bootstrap_crosswalk]{bootstrap_crosswalk()}}. See
\code{\link[=boot_control]{boot_control()}} for more information.}

\item{method}{Either 'lm' (recommended) or 'manual'. The former will fit a linear
regression model and return the fit object, while the latter will return a table
containing the covariance between the cognitive measures (\code{cov}), the variance of
the measure input as cog1 (\code{var}), and the estimated sloped (\code{coef}).}
}
\description{
Crosswalk cognitive measures
}
\details{
Unlike \code{\link[=crosswalk]{crosswalk()}}, which uses splitting to estimate the sample coefficient,
\code{\link[=cw]{cw()}} calculates the coefficient based on summary statistics from the input
data: cov(cog1, cog2) / var(cog1).
}
