% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{plot.cogxwalkr}
\alias{plot.cogxwalkr}
\title{Plot information about the bootstrap distribution}
\usage{
\method{plot}{cogxwalkr}(
  cx,
  cxsum = NULL,
  types = c("boot", "slope"),
  breaks = "FD",
  citype = "percentile",
  layout = c(1, length(types)),
  sargs = list(col = "black", lty = 1, lwd = 2),
  bargs = list(col = "red", lty = 2, lwd = 2),
  slargs = list(col = "red", lwd = 2),
  clargs = list(col = "red", lty = 2),
  ptshape = 19,
  ptsize = 0.8,
  ptcol = "black",
  ptalpha = 0.2
)
}
\arguments{
\item{cx}{An object of class "cogxwalkr", i.e., an object returned by \code{\link[=crosswalk]{crosswalk()}}.}

\item{cxsum}{The output of \code{summary(cx)}}

\item{types}{The types of crosswalk plots to produce. By default, both a plot of the
bootstrap distribution of coefficients and a plot of the data with the estimated
slope. For the plotted sloped, (1-alpha)\% confidence intervals will appear if the
user provides the output of \code{\link[=summary.cogxwalkr]{summary.cogxwalkr()}}.}

\item{breaks}{Passed to \code{\link[graphics:hist]{graphics::hist()}}, overriding the default method with "FD"
(\code{\link[grDevices:nclass]{grDevices::nclass.FD()}}).}

\item{citype}{Choose confidence intervals to plot. Ignored if \code{cxsum} is NULL.}

\item{layout}{Passed to the \code{mfrow} argument of \code{\link[graphics:par]{graphics::par()}}. Defaults to
\code{c(1, length(types))}.}

\item{sargs}{List of parameters passed to the \code{\link[graphics:abline]{graphics::abline()}} that plots the
sample coefficient estimate}

\item{bargs}{List of parameters passed to the \code{\link[graphics:abline]{graphics::abline()}} that plots the mean
coefficient estimate across bootstrap replicates}

\item{slargs}{List of parameters passed to the \code{\link[graphics:abline]{graphics::abline()}} that plots the
sample slope in the crosswalk panel}

\item{clargs}{List of parameters passed to the \code{\link[graphics:abline]{graphics::abline()}} calls that plot
the confidence limits for the slope.}

\item{ptshape}{Shape of points in crosswalk scatterplot (passed to \code{\link[base:plot]{base::plot()}})}

\item{ptsize}{Size of points in crosswalk scatterplot (passed to \code{\link[base:plot]{base::plot()}})}

\item{ptcol}{Color of points in crosswalk scatterplot (passed to \code{\link[base:plot]{base::plot()}})}

\item{ptalpha}{Alpha (transparency) of points in crosswalk scatterplot
(passed to \code{\link[base:plot]{base::plot()}})}
}
\description{
Plot information about the bootstrap distribution
}
