% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{print.summary.cogxwalkr}
\alias{print.summary.cogxwalkr}
\title{Print a cogxwalkr summary}
\usage{
\method{print}{summary.cogxwalkr}(x, digits = 3L)
}
\arguments{
\item{digits}{Number of digits to print. Passed to \code{round()}.}
}
\description{
Print a cogxwalkr summary
}
