% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.r
\name{make_unconditional_splits}
\alias{make_unconditional_splits}
\alias{make_conditional_splits}
\alias{make_splits}
\title{Splitting methods}
\usage{
make_unconditional_splits(data, niter = NULL)

make_conditional_splits(cdvar = NULL, data, loop = FALSE)

make_splits(cdvar = NULL, data, cdloop = FALSE, niter = NULL)
}
\arguments{
\item{data}{A data.table, data.frame, matrix, or list containing the
cognitive measure data}

\item{niter}{Number of iterations to conduct for an unconditional split routine}

\item{cdvar}{Character string naming auxiliary variable by which to condition splits}

\item{loop}{Boolean declaring whether to use a for loop. The default FALSE will
generate splits and operate on an expanded data.table in-memory. If your machine has
limited memory, set this argument to TRUE in order to process splits sequentially.
The default option should be much faster.}

\item{cdloop}{Boolean passed to \code{loop} argument of \code{make_conditional_splits()}. Ignored
when conducting unconditional splits.}
}
\description{
These functions handle split routines used in the estimation of crosswalks. In general,
we recommend using these functions via \code{\link[=crosswalk]{crosswalk()}}, but they are made available to
the user who wishes to call them directly.
\itemize{
\item \code{\link[=make_unconditional_splits]{make_unconditional_splits()}} : implements the unconditional split routine
\item \code{\link[=make_conditional_splits]{make_conditional_splits()}} : implements the conditional split routine, in which
splits in the data are defined by a binary variable (e.g., dementia) correlated
with given cognitive measures (e.g., MMSE and MoCA) because the underlying
cognitive construct is a common cause of all three
\item \code{\link[=make_splits]{make_splits()}} : a handler that checks the input dataset and passes
it on to the appropriate split function with relevant arguments specified
}
}
