% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.r
\name{summary.cogxwalkr}
\alias{summary.cogxwalkr}
\title{Summarize a cogxwalkr list}
\usage{
\method{summary}{cogxwalkr}(cx, alpha = 0.05, bci_type)
}
\arguments{
\item{cx}{An object of class "cogxwalkr", i.e., an object returned by \code{crosswalk()}.}

\item{alpha}{Alpha to use for confidence interval calculation. Defaults to 0.05.}

\item{bci_type}{Type of bootstrapped confidence interval to calculate. Currently
only "percentile" is supported.}
}
\description{
Summarize a cogxwalkr list
}
